chore : 리뷰 반영 코드 위치 수정 / 패키지 이름 변경
