--- src/main/java/in/koreatech/koin/domain/TimeTable/controller/SemesterApi.java
@@ -1,30 +1,0 @@
-package in.koreatech.koin.domain.TimeTable.controller;
-
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-
-import in.koreatech.koin.domain.TimeTable.dto.SemesterResponse;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.tags.Tag;
-
-@Tag(name = "(Normal) TimeTable : 시간표", description = "시간표 관련 API")
-public interface SemesterApi {
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "401", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true)))
-        }
-    )
-    @Operation(summary = "학기 정보 조회")
-    @GetMapping("/semesters")
-    ResponseEntity<List<SemesterResponse>> getSemesters();
-
-}
