+++ src/main/java/in/koreatech/koin/domain/timetable/controller/SemesterController.java
@@ -1,0 +1,14 @@
+package in.koreatech.koin.domain.timetable.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequiredArgsConstructor
+public class SemesterController {
+
+
+
+
+}
